ps-120 Added product page component and refactored code for bestsellers
