+++ web/src/components/pages/index.ts
@@ -4,4 +4,3 @@ export * from './cartPage/CartPage';
 export * from './sportsPage/SportsPage';
 export * from './singleProductPage/SingleProductPage';
 export * from './brandsPage/BrandsPage';
-export * from './bestellersPage/BestsellersPage';
