--- web/src/constants.ts
@@ -1,1 +1,0 @@
-export const CURRENCY_TYPE = '$';
