+++ web/src/constants/index.ts
@@ -1,1 +1,3 @@
-export {}
+export const CURRENCY_TYPE = '$';
+export const ITEMS_ON_PAGE = 12;
+export const BESTSELLER_SELL_COUNT = 20;
