+++ web/src/routes.tsx
@@ -5,7 +5,6 @@ import {
     SingleProductPage,
     SportsPage,
     BrandsPage,
-    BestsellersPage,
 } from './components';
 import { RoutePaths } from './models';
 
@@ -50,11 +49,5 @@ export const routes = [
         exact: true,
         element: <BrandsPage />,
         key: 'brands-page',
-    },
-    {
-        path: RoutePaths.Bestsellers,
-        exact: true,
-        element: <BestsellersPage />,
-        key: 'bestsellers-page',
     },
 ];
