+++ web/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './randomId';
 export * from './filterArrayByType';
+export * from './filterArrayBySellCount';
